#10413: minor refactoring
